Move header layout to shared folder
