+++ web/src/app/components/sign-in-page/sign-in-page.module.ts
@@ -16,7 +16,7 @@
 
 import { BrowserModule } from '@angular/platform-browser';
 import { CommonModule } from '@angular/common';
-import { HeaderLayoutModule } from 'app/components/header-layout/header-layout.module';
+import { HeaderLayoutModule } from 'app/shared/components/header-layout/header-layout.module';
 import { FirebaseUIModule } from 'firebaseui-angular';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { MatButtonModule } from '@angular/material/button';
