+++ web/src/app/shared/components/header-layout/current-user-widget/current-user-widget.component.spec.ts
