+++ web/src/app/shared/components/header-layout/header-layout.component.scss
